Separation of effect handling from Entity, fixes #886
